livestock/requirements.txt
pillow
phone_field
django cleanup
